Added full javadoc to MessageHistory. Moved MessageHistory to the entities package. Removed static getHistoryAround from MessageHistory, now lives only as a default method in MessageChannel. Added target="_blank" to some href links. Cleaned up a few javadoc errors.
