+++ src/main/java/net/dv8tion/jda/core/EmbedBuilder.java
@@ -27,7 +27,8 @@ import java.util.regex.Pattern;
 
 /**
  * Builder system used to build {@link net.dv8tion.jda.core.entities.MessageEmbed MessageEmbeds}.
- * <br>A visual breakdown of an Embed and how it relates to this class is available <a href="http://imgur.com/a/yOb5n">here</a>.
+ * <br>A visual breakdown of an Embed and how it relates to this class is available at
+ * <a href="http://imgur.com/a/yOb5n" target="_blank">http://imgur.com/a/yOb5n</a>.
  *
  * @since  JDA 3.0
  * @author John A. Grosh
